+++ modules/actor-adaptor-std/Cargo.toml
@@ -20,7 +20,7 @@ tokio-executor = ["dep:tokio"]
 [dependencies]
 fraktor-actor-core-rs = { workspace = true }
 fraktor-utils-adaptor-std-rs = { workspace = true }
-fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize"] }
+fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize", "std-locks"] }
 tokio = { workspace = true, optional = true, features = ["rt", "rt-multi-thread", "macros", "time"] }
 tracing = { workspace = true, features = ["std"] }
