+++ modules/cluster-adaptor-std/Cargo.toml
@@ -21,7 +21,7 @@ test-support = ["fraktor-cluster-core-rs/test-support", "fraktor-actor-core-rs/t
 fraktor-cluster-core-rs = { workspace = true }
 fraktor-actor-core-rs = { workspace = true }
 fraktor-actor-adaptor-std-rs = { workspace = true }
-fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize"] }
+fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize", "std-locks"] }
 serde = { workspace = true }
 postcard = { workspace = true }
 tracing = { workspace = true, features = ["std"] }
