+++ modules/remote-adaptor-std/Cargo.toml
@@ -23,7 +23,7 @@ test-support = [
 fraktor-remote-core-rs = { workspace = true }
 fraktor-actor-core-rs = { workspace = true }
 fraktor-actor-adaptor-std-rs = { workspace = true }
-fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize"] }
+fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize", "std-locks"] }
 bytes = { workspace = true }
 futures = { workspace = true, features = ["std"] }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "net", "sync", "time", "io-util"] }
