+++ modules/utils-adaptor-std/Cargo.toml
@@ -16,7 +16,7 @@ autoexamples = false
 default = []
 
 [dependencies]
-fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize"] }
+fraktor-utils-core-rs = { workspace = true, features = ["alloc", "unsize", "std-locks"] }
 spin = { workspace = true, default-features = false, features = ["mutex", "spin_mutex", "rwlock", "portable_atomic"] }
 
 [dev-dependencies]
