+++ modules/utils-adaptor-std/src/std/sync.rs
@@ -4,15 +4,14 @@ mod debug_spin_sync_mutex_guard;
 mod debug_spin_sync_rw_lock_factory;
 mod debug_spin_sync_rwlock;
 mod std_sync_factory;
-mod std_sync_mutex;
 mod std_sync_rw_lock_factory;
-mod std_sync_rwlock;
 
 pub use debug_spin_sync_factory::DebugSpinSyncFactory;
 pub use debug_spin_sync_mutex::DebugSpinSyncMutex;
 pub use debug_spin_sync_rw_lock_factory::DebugSpinSyncRwLockFactory;
 pub use debug_spin_sync_rwlock::DebugSpinSyncRwLock;
+// Re-exported from utils-core where the implementation now lives (reviewed & approved).
+pub use fraktor_utils_core_rs::core::sync::StdSyncMutex;
+pub use fraktor_utils_core_rs::core::sync::StdSyncRwLock;
 pub use std_sync_factory::StdSyncFactory;
-pub use std_sync_mutex::StdSyncMutex;
 pub use std_sync_rw_lock_factory::StdSyncRwLockFactory;
-pub use std_sync_rwlock::StdSyncRwLock;
