+++ modules/utils-adaptor-std/src/std/sync/std_sync_rw_lock_factory.rs
@@ -1,6 +1,6 @@
 use fraktor_utils_core_rs::core::sync::RwLockDriverFactory;
 
-use super::std_sync_rwlock::StdSyncRwLock;
+use super::StdSyncRwLock;
 
 /// Factory for [`StdSyncRwLock`](super::StdSyncRwLock).
 pub struct StdSyncRwLockFactory;
