+++ modules/utils-core/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["concurrency", "embedded", "no-std"]
 default = ["alloc", "unsize"]
 alloc = []
 std = []
+std-locks = ["std"]
 debug-locks = ["std"]
 interrupt-cortex-m = ["cortex-m"]
 force-portable-arc = []
